Fix Refactor Issues
