+++ components/org.wso2.identity.webhook.wso2.event.handler/src/main/java/org/wso2/identity/webhook/wso2/event/handler/internal/constant/Constants.java
@@ -32,4 +32,5 @@ public class Constants {
     public static final String CURRENT_AUTHENTICATOR_ERROR_CODE = "currentAuthenticatorErrorCode";
     public static final String DATA_MAP = "dataMap";
     public static final String WSO2_EVENT_SCHEMA = "WSO2";
+    public static final String ORGANIZATION_AUTHENTICATOR = "OrganizationAuthenticator";
 }
